+++ UniKP/pretrain_trfm.py
@@ -12,8 +12,8 @@ from torch.nn import functional as F
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
-from build_vocab import WordVocab
-from dataset import Seq2seqDataset
+from UniKP.build_vocab import WordVocab
+from UniKP.dataset import Seq2seqDataset
 
 PAD = 0
 UNK = 1
