+++ scripts/UniKP_kcat.py
@@ -1,28 +1,35 @@
+from UniKP.utils import Seq_to_vec, smiles_to_vec
 import json
-
 from sklearn.ensemble import ExtraTreesRegressor
 import numpy as np
 import pandas as pd
-from sklearn.model_selection import KFold
+
+import random
 import pickle
 import math
 
 
-def Kcat_predict(Ifeature, Label, weights):
-    kf = KFold(n_splits=5, shuffle=True)
-    All_pre_label = []
-    All_real_label = []
-    for train_index, test_index in kf.split(Ifeature, Label):
+
+def Kcat_predict(Ifeature, Label, sequence_new, Smiles_new, ECNumber_new, Organism_new, Substrate_new, Type_new):
+    for i in range(10):
+        # Generate training or test set index
+        ALL_index = [j for j in range(len(Ifeature))]
+        train_index = np.array(random.sample(ALL_index, int(len(ALL_index)*0.9)))
+        Training_or_test = []
+        for j in range(len(ALL_index)):
+            if ALL_index[j] in train_index:
+                Training_or_test.append(0)
+            else:
+                Training_or_test.append(1)
+        Training_or_test = np.array(Training_or_test)
         Train_data, Train_label = Ifeature[train_index], Label[train_index]
-        Test_data, Test_label = Ifeature[test_index], Label[test_index]
         model = ExtraTreesRegressor()
-        # , sample_weight=weights[train_index]
-        model.fit(Train_data, Train_label, sample_weight=weights[train_index])
-        Pre_label = model.predict(Test_data)
-        All_pre_label.extend(Pre_label)
-        All_real_label.extend(Test_label)
-    res = pd.DataFrame({'Value': All_real_label, 'Predict_Label': All_pre_label})
-    res.to_excel('DMW/DMW_No_Normalize_2_LDS_Kcat_5_cv'+'.xlsx')
+        model.fit(Train_data, Train_label)
+        Pre_all_label = model.predict(Ifeature)
+        res = pd.DataFrame({'sequence': sequence_new, 'smiles': Smiles_new, 'ECNumber': ECNumber_new,
+                            'Organism': Organism_new, 'Substrate': Substrate_new, 'Type': Type_new,
+                            'Label': Label, 'Predict_Label': Pre_all_label, 'Training or test': Training_or_test})
+        res.to_excel('PreKcat_new/'+str(i+1)+'_all_samples_metrics.xlsx')
 
 
 if __name__ == '__main__':
@@ -30,7 +37,6 @@ if __name__ == '__main__':
     with open('Kcat_combination_0918_wildtype_mutant.json', 'r') as file:
         datasets = json.load(file)
     # print(len(datasets))
-    # datasets = datasets[:50]
     sequence = [data['Sequence'] for data in datasets]
     Smiles = [data['Smiles'] for data in datasets]
     Label = [float(data['Value']) for data in datasets]
@@ -43,14 +49,14 @@ if __name__ == '__main__':
             Label[i] = -10000000000
         else:
             Label[i] = math.log(Label[i], 10)
-    Label = np.array(Label)
     print(max(Label), min(Label))
     # Feature Extractor
-    # smiles_input = smiles_to_vec(Smiles)
-    # sequence_input = Seq_to_vec(sequence)
-    # feature = np.concatenate((smiles_input, sequence_input), axis=1)
-    with open("PreKcat_new/features_17010_PreKcat.pkl", "rb") as f:
-        feature = pickle.load(f)
+    smiles_input = smiles_to_vec(Smiles)
+    sequence_input = Seq_to_vec(sequence)
+    feature = np.concatenate((smiles_input, sequence_input), axis=1)
+    with open("PreKcat_new/features_16838_PreKcat.pkl", "wb") as f:
+        pickle.dump(feature, f)
+    Label = np.array(Label)
     # Input dataset
     feature_new = []
     Label_new = []
@@ -71,16 +77,8 @@ if __name__ == '__main__':
             Substrate_new.append(Substrate[i])
             Type_new.append(Type[i])
     print(len(Label_new), min(Label_new), max(Label_new))
-    feature_new = np.array(feature_new)
     Label_new = np.array(Label_new)
-    # sl_label = [Label_new[i] for i in range(len(Label_new))]
-    # weights = Smooth_Label(sl_label)
-    weights = np.ones([len(Label_new)], dtype=float)
-    for i in range(len(weights)):
-        if Label_new[i] > 4:
-            weights[i] = 2
-    # sum_weights = np.sum(weights)
-    # for i in range(len(weights)):
-    #     weights[i] /= sum_weights
+    feature_new = np.array(feature_new)
     # Modelling
-    Kcat_predict(feature_new, Label_new, weights)
+    Kcat_predict(feature_new, Label_new, sequence_new, Smiles_new, ECNumber_new,
+                Organism_new, Substrate_new, Type_new)
