+++ scripts/smi2csv.py
